Calculate relative URLs to edit-plugin and add-plugin views correctly.
This is important if the form for editing the plugin has any links
to other pages (e.g. if the plugin has a foreign key to another model,
and there is a select box with a '+' sign for adding more).  Otherwise
the URLs to other views won't be calculated properly.

Signed-off-by: Patrick Lauber <patrick.lauber@divio.ch>