Add test to check LogEntry added when pages edited via AJAX
This happens when using the page list - if you (un)check the checkboxes
for "published" or "in navigation" an AJAX request happens to do this
immediately.  However, until now, this has not added a LogEntry in the
way that the admin system does when you save an object in the admin
system.