Added stopPropagation() to dblclick handler
This is to address the issue of not being able to edit models emitted by CMSPlugins. Because the event was being propagated, the double-click essentially triggers editing of both the model and of the plugin. The resulting mayhem made it impossible to do either. This fix seems to allow the editing of the model, without preventing edits to the plugin (by double-clicking on something other than the model).

Signed-off-by: Martin Koistinen <mkoistinen@gmail.com>