If the language menu isn't present, don't change it.
As per PR #2651, the language menu was disabled (not rendered) if the settings.USE_I18N is False. Previously, code would alter that menu, whether or not it was present, resulting in a mis-formed language menu.

This PR adds a new public API for the toolbar get_menu(), which operates like get_or_create_menu(), except that it will return None if the menu is not found.

Docs also updated.

Tests pass.