Updated docs
I added a warning about returning non-utf8-encoded strings from within ``def __unicode__()``, as I found it resulting in a 500 response and a plugin instance marked as <Empty>. After returning as UTF8 everything worked as expected. If you can't confirm these issues to be related or feel that the addition doesn't make sense for documentation readers, feel free to drop it.