change the timing on user settings sideframe reload
when the page is loaded in sideframe we want to override the url of the
sideframe and reload the parent so the change in language is reflected
in the UI. however there's also a load handler for the sideframe to save
the url in the settings so that after you reload - correct frame is
restored. problem is, this load event happens after we try to modify the
sideframe url manually, so we want ours to trigger last.