Move urls.SLUG_REGEXP to constants
This fixes the problem with Django 1.9 system checks. In a few words:
importing from cms.urls triggers app reseolver mechanism that fails
if there were no database tables for contrib.sites created yet.
It catches `DoesNotExist` but not `OperationalError`. Moving this
constant out to a different localtion seems to be a better fix.