Fixed a regression when standalone widgets wouldn't work
The regression happened because in
https://github.com/divio/django-cms/commit/a44cf25d6d5fd63cd2e9bd2ad1b5e14ae15c7d7b
and before that in 3.3.0 the files required for widgets were
removed from media js class on the assumption that the CMS provided
bundles would be loaded anyway. This commit changes the way
requirements are loaded for those widgets so they can be used
standalone and on CMS pages as well.