feat: backport - django-cms 4.0.x - Django 3.0 support (#7105)
Based on changes made in the following PR: 6c92fbe#diff-069df57a3a645590c8bd00dfffb2bd94969c691857a20c7936a5e06fa122a8f3

This is a backport of the original change for django 3.0. This has been done to have a direct comparison to the original change. There will be another PR for django 3.1 and another for 3.2 which is a few PRs combined.